Add support for multiple transforms
The second (example) transform was added. If configured, it makes all
fields in all classes public.

Right now both transforms could not be applied - the second one always
fails on some class in the middle of the processing, idnependent of
which transform was first and which is the second. For this reason the
FieldExposer transform is turned off for the time being.