Add the field access modifier - a really dirty hack
This allows to change the class field access including static/non static
changes. Fields without initializer or with simple initializer are
supported. The simple initializer is initializer with constant or with a
single static call. See test for some examples.

Note: fields with various initializers in different constructors are not
supported as well as some complicated initialization. I.e. the code like
this:

    class SomeClass {
        int counter;

        SomeClass() {
            counter = 0;
        }

        SomeClass(int initialValue) {
            counter = initialValue;
        }
    }

or like this:

    class SomeClass {
        static String text;
        static {
            long time = System.currentTimeMillis();
            text = "The class loading time is " + time;
        }
    }

is a really bad example for this transformation. If you will lucky, the
transformer will fail with a readable description of the problem.
Otherwise you will face with the old bad undefined behavior.